cloud_function_jira_processor/requirements.txt
functions-framework==3.*
cloudevents==1.10.1
google-cloud-secret-manager==2.16.4
jira==3.5.2
